+++ lib/arjdbc/as400/adapter.rb
@@ -81,8 +81,8 @@ module ArJdbc
     def execute_and_auto_confirm(sql, name = nil)
 
       begin
-        @connection.execute_update "CALL qsys.qcmdexc('QSYS/CHGJOB INQMSGRPY(*SYSRPYL)', 0000000031.00000)"
-        @connection.execute_update "CALL qsys.qcmdexc('ADDRPYLE SEQNBR(9876) MSGID(CPA32B2) RPY(''I'')', 0000000045.00000)"
+        @connection.execute_system_command('QSYS/CHGJOB INQMSGRPY(*SYSRPYL)')
+        @connection.execute_system_command("ADDRPYLE SEQNBR(9876) MSGID(CPA32B2) RPY('I')")
       rescue Exception => e
         raise "Could not call CHGJOB INQMSGRPY(*SYSRPYL) and ADDRPYLE SEQNBR(9876) MSGID(CPA32B2) RPY('I').\n" +
               "Do you have authority to do this?\n\n#{e.inspect}"
@@ -99,8 +99,8 @@ module ArJdbc
 
         # Ensure default configuration restoration
         begin
-          @connection.execute_update "CALL qsys.qcmdexc('QSYS/CHGJOB INQMSGRPY(*DFT)', 0000000027.00000)"
-          @connection.execute_update "CALL qsys.qcmdexc('RMVRPYLE SEQNBR(9876)', 0000000021.00000)"
+          @connection.execute_system_command('QSYS/CHGJOB INQMSGRPY(*DFT)')
+          @connection.execute_system_command('RMVRPYLE SEQNBR(9876)')
         rescue Exception => e
           raise "Could not call CHGJOB INQMSGRPY(*DFT) and RMVRPYLE SEQNBR(9876).\n" +
                     "Do you have authority to do this?\n\n#{e.inspect}"
@@ -133,6 +133,32 @@ module ArJdbc
       db2_schema
     end
 
+    def execute_system_command(command)
+      command = ::ActiveRecord::Base.sanitize(command)
+      @connection.execute_update("CALL qsys.qcmdexc(#{command}, CAST(#{command.length - 2} AS DECIMAL(15, 5)))")
+    end
+
+    # Disable transactions when they are not supported
+    def begin_db_transaction
+      super unless config[:transaction_isolation] == 'none'
+    end
+
+    def commit_db_transaction
+      super unless config[:transaction_isolation] == 'none'
+    end
+
+    def rollback_db_transaction
+      super unless config[:transaction_isolation] == 'none'
+    end
+
+    def begin_isolated_db_transaction(isolation)
+      super unless config[:transaction_isolation] == 'none'
+    end
+
+    def create_savepoint(name = current_savepoint_name(true))
+      super unless config[:transaction_isolation] == 'none'
+    end
+
     private
     # If naming is really in system mode CURRENT_SCHEMA is *LIBL
     def system_naming?
@@ -152,10 +178,9 @@ module ArJdbc
           config[:schema]
         else
           # Only found method to set db2_schema from jndi
-          result = select_one("SELECT CURRENT_SCHEMA FROM SYSIBM.SYSDUMMY1")
+          result = select_one('VALUES CURRENT_SCHEMA')
           result['00001']
         end
     end
-
   end
 end